+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Gear mod kit, which can remove some types of parts from gear items
+- Very crude repair kit, for those who struggle to find iron
+
 ## [2.3.2] - 2020-10-22
 ### Added
 - Basalt material